move methods for getting fills or orders
