+++ broker-daemon/models/block-order.spec.js
@@ -995,110 +995,6 @@ describe('BlockOrder', () => {
     })
   })
 
-  describe('::getOrdersForRange', () => {
-    let orders
-    let getRecords
-    let orderStub
-    let storeStub
-    let parse
-
-    const reverts = []
-
-    beforeEach(() => {
-      orders = [
-        { id: 1234 },
-        { id: 1337 }
-      ]
-      getRecords = sinon.stub().resolves(orders)
-      orderStub = {
-        rangeForIds: sinon.stub(),
-        fromObject: sinon.stub()
-      }
-      storeStub = sinon.stub()
-      parse = sinon.stub().returns({})
-
-      reverts.push(BlockOrder.__set__('getRecords', getRecords))
-      reverts.push(BlockOrder.__set__('Order', orderStub))
-      reverts.push(BlockOrder.__set__('JSON', { parse }))
-    })
-
-    afterEach(() => {
-      reverts.forEach(r => r())
-    })
-
-    it('gets records for a specific store', async () => {
-      await BlockOrder.getOrdersForRange(storeStub, orders[0].id, orders[1].id)
-      expect(getRecords).to.have.been.calledWith(storeStub, sinon.match.func, sinon.match.any)
-    })
-
-    it('inflates an order', async () => {
-      await BlockOrder.getOrdersForRange(storeStub, orders[0].id, orders[1].id)
-      const recordFilter = getRecords.args[0][1]
-      const value = '{}'
-      const key = 'key'
-      recordFilter(key, value)
-      expect(parse).to.have.been.calledWith(value)
-      expect(orderStub.fromObject).to.have.been.calledWith(key, sinon.match.any)
-    })
-
-    it('gets records for a particular range', async () => {
-      await BlockOrder.getOrdersForRange(storeStub, orders[0].id, orders[1].id)
-      expect(orderStub.rangeForIds).to.have.been.calledWith(orders[0].id, orders[1].id)
-    })
-  })
-
-  describe('::getFillsForRange', () => {
-    let fills
-    let getRecords
-    let fillStub
-    let storeStub
-    let parse
-
-    const reverts = []
-
-    beforeEach(() => {
-      fills = [
-        { id: 1234 },
-        { id: 1337 }
-      ]
-      getRecords = sinon.stub().resolves(fills)
-      fillStub = {
-        rangeForIds: sinon.stub(),
-        fromObject: sinon.stub()
-      }
-      storeStub = sinon.stub()
-      parse = sinon.stub().returns({})
-
-      reverts.push(BlockOrder.__set__('getRecords', getRecords))
-      reverts.push(BlockOrder.__set__('Fill', fillStub))
-      reverts.push(BlockOrder.__set__('JSON', { parse }))
-    })
-
-    afterEach(() => {
-      reverts.forEach(r => r())
-    })
-
-    it('gets records for a specific store', async () => {
-      await BlockOrder.getFillsForRange(storeStub, fills[0].id, fills[1].id)
-      expect(getRecords).to.have.been.calledWith(storeStub, sinon.match.func, sinon.match.any)
-    })
-
-    it('inflates a fill', async () => {
-      await BlockOrder.getFillsForRange(storeStub, fills[0].id, fills[1].id)
-      const recordFilter = getRecords.args[0][1]
-      const value = '{}'
-      const key = 'key'
-      recordFilter(key, value)
-      expect(parse).to.have.been.calledWith(value)
-      expect(fillStub.fromObject).to.have.been.calledWith(key, sinon.match.any)
-    })
-
-    it('gets records for a particular range', async () => {
-      await BlockOrder.getFillsForRange(storeStub, fills[0].id, fills[1].id)
-      expect(fillStub.rangeForIds).to.have.been.calledWith(fills[0].id, fills[1].id)
-    })
-  })
-
   describe('::activeAmountsForOrders', () => {
     const OrderStateMachine = BlockOrder.__get__('OrderStateMachine')
     const { CREATED, PLACED, EXECUTING } = OrderStateMachine.STATES
