+++ broker-daemon/models/order.js
@@ -1,4 +1,7 @@
-const { Big } = require('../utils')
+const {
+  Big,
+  getRecords
+} = require('../utils')
 const CONFIG = require('../config')
 /**
  * Delimiter for the block order id and order when storing orders
@@ -473,17 +476,23 @@ class Order {
   }
 
   /**
-   * Create a set of options that can be passed to a Level `createReadStream` call
-   * that limits the set to orders that belong to the given range of blockOrderIds.
-   * @param {string} firstId - blockOrderId of the start of our block order range
-   * @param {string} lastId - blockOrderId of the end of our block order range
-   * @returns {Object} Options object that can be used in {@link https://github.com/Level/levelup#createReadStream}
+   * Grabs all orders for a range of block orders
+   * @param {sublevel} store
+   * @param {string} first
+   * @param {string} last
+   * @returns {void}
    */
-  static rangeForIds (firstId, lastId) {
-    return {
-      gte: `${firstId}${DELIMITER}${LOWER_BOUND}`,
-      lte: `${lastId}${DELIMITER}${UPPER_BOUND}`
-    }
+  static async getOrdersForRange (store, first, last) {
+    return getRecords(
+      store,
+      (key, value) => {
+        const { order, state, error, dates } = JSON.parse(value)
+        return { order: Order.fromObject(key, order), state, error, dates }
+      },
+      // limit the orders we retrieve to those that belong to this blockOrder, i.e. those that are in
+      // its prefix range.
+      Order.rangeForBlockOrderIds(first, last)
+    )
   }
 }
